requirements.txt
uvicorn
fastapi
starlette
uvicorn
fastapi-multipart
tensorflow
Pillow
numpy
